flask-finance-chatbot/app/requirements/base.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
openai==0.28.0
requests==2.31.0
python-jose==3.3.0
werkzeug==2.3.7 